Dev query4 (#9)
* Fix: removed cast extra data now works (removed invalid rows). Boundary sends correct credits data

* Fix: worker import typo and removed unused method from client

* Update: boundary can receive query 3 and write it to a file

* Update: sending ratings_small.csv

* Update: sending ratings_small.csv

* Add: avg movies rating done.
TODO: implement sharding

* Update: pushing .env for convenience. No sensitive data is being exposed so its ok

* Add: join rating worker. It does not process various clients, only one join per client.
TODO: change this so we do a fanout. All workers process every client (no problem in every worker having a state for each active client) so they can distribute the ratings load between joiners

* Add: requirements.txt for joiner

* Add: max/min worker.
TODO: shard it and add the collector worker at the end so it collects all the responses and pushes it to the response queue (see professor's DAG for reference)

* Update: docker compose for max/min query 3. WARNING: Commented out the query 4 and 5.

* Add fanout to join routers and remove some debug prints.
All queries (except Q2) integrated and working.

---------

Co-authored-by: ignacioregazzoli <ignacioregazzoli@gmail.com>